+++ guillotina/interfaces/content.py
@@ -284,6 +284,16 @@ class IGetOwner(Interface):
     Defines a utility for calculating the owner of a new resource
     """
 
+    def __init__(context):
+        """
+        adapts context
+        """
+
+    async def __call__(default):
+        """
+        get owner
+        """
+
 
 class IIDGenerator(Interface):
     """